+++ bofire/data_models/kernels/api.py
@@ -1,5 +1,3 @@
-from typing import Union
-
 from bofire.data_models.kernels._register import register_kernel  # noqa: F401
 from bofire.data_models.kernels.aggregation import (
     AdditiveKernel,
@@ -30,16 +28,8 @@ from bofire.data_models.kernels.kernel import (
 )
 from bofire.data_models.kernels.molecular import MolecularKernel, TanimotoKernel
 from bofire.data_models.kernels.shape import WassersteinKernel
+from bofire.data_models.unions import tagged_union
 
-
-AbstractKernel = Union[
-    Kernel,
-    CategoricalKernel,
-    ContinuousKernel,
-    MolecularKernel,
-    FeatureSpecificKernel,
-    AggregationKernel,
-]
 
 _CONTINUOUS_KERNEL_TYPES: list[type[ContinuousKernel]] = [
     MaternKernel,
@@ -50,7 +40,7 @@ _CONTINUOUS_KERNEL_TYPES: list[type[ContinuousKernel]] = [
     InfiniteWidthBNNKernel,
 ]
 
-AnyContinuousKernel = Union[tuple(_CONTINUOUS_KERNEL_TYPES)]
+AnyContinuousKernel = tagged_union(*_CONTINUOUS_KERNEL_TYPES)
 
 _CATEGORICAL_KERNEL_TYPES: list[type[CategoricalKernel]] = [
     HammingDistanceKernel,
@@ -58,7 +48,7 @@ _CATEGORICAL_KERNEL_TYPES: list[type[CategoricalKernel]] = [
     PositiveIndexKernel,
 ]
 
-AnyCategoricalKernel = Union[tuple(_CATEGORICAL_KERNEL_TYPES)]
+AnyCategoricalKernel = tagged_union(*_CATEGORICAL_KERNEL_TYPES)
 
 AnyMolecularKernel = TanimotoKernel
 
@@ -81,4 +71,4 @@ _KERNEL_TYPES: list[type[Kernel]] = [
     WedgeKernel,
 ]
 
-AnyKernel = Union[tuple(_KERNEL_TYPES)]
+AnyKernel = tagged_union(*_KERNEL_TYPES)