+++ bofire/data_models/strategies/actual_strategy_type.py
@@ -1,5 +1,3 @@
-from typing import Union
-
 from bofire.data_models.strategies.doe import DoEStrategy
 from bofire.data_models.strategies.factorial import FactorialStrategy
 from bofire.data_models.strategies.fractional_factorial import (
@@ -25,6 +23,7 @@ from bofire.data_models.strategies.predictives.sobo import (
 from bofire.data_models.strategies.random import RandomStrategy
 from bofire.data_models.strategies.shortest_path import ShortestPathStrategy
 from bofire.data_models.strategies.strategy import Strategy
+from bofire.data_models.unions import tagged_union
 
 
 _ACTUAL_STRATEGY_TYPES: list[type[Strategy]] = [
@@ -46,4 +45,4 @@ _ACTUAL_STRATEGY_TYPES: list[type[Strategy]] = [
     LLMStrategy,
 ]
 
-ActualStrategy = Union[tuple(_ACTUAL_STRATEGY_TYPES)]
+ActualStrategy = tagged_union(*_ACTUAL_STRATEGY_TYPES)