+++ bofire/data_models/strategies/doe.py
@@ -1,4 +1,4 @@
-from typing import Annotated, Dict, List, Literal, Optional, Type, Union
+from typing import Annotated, Any, Dict, List, Literal, Optional, Type, Union
 
 from formulaic import Formula
 from formulaic.errors import FormulaSyntaxError
@@ -10,6 +10,7 @@ from bofire.data_models.features.api import Feature
 from bofire.data_models.objectives.api import Objective
 from bofire.data_models.strategies.strategy import Strategy
 from bofire.data_models.types import Bounds
+from bofire.data_models.unions import tagged_union
 
 
 PREDEFINED_MODEL_TYPES = Literal[
@@ -21,7 +22,7 @@ PREDEFINED_MODEL_TYPES = Literal[
 
 
 class OptimalityCriterion(BaseModel):
-    type: str
+    type: Any
     delta: float = 1e-6
     transform_range: Optional[Bounds] = None
 
@@ -32,7 +33,7 @@ class SpaceFillingCriterion(OptimalityCriterion):
 
 
 class DoEOptimalityCriterion(OptimalityCriterion):
-    type: str
+    type: Any
     formula: Union[PREDEFINED_MODEL_TYPES, str]
     """
     model_type (str, Formula): keyword or formulaic Formula describing the model. Known keywords
@@ -77,19 +78,24 @@ class IOptimalityCriterion(DoEOptimalityCriterion):
     ipopt_options: Optional[Dict] = None
 
 
-AnyDoEOptimalityCriterion = Union[
+AnyDoEOptimalityCriterion = tagged_union(
     KOptimalityCriterion,
     GOptimalityCriterion,
     AOptimalityCriterion,
     EOptimalityCriterion,
     DOptimalityCriterion,
     IOptimalityCriterion,
-]
+)
 
-AnyOptimalityCriterion = Union[
-    AnyDoEOptimalityCriterion,
+AnyOptimalityCriterion = tagged_union(
+    KOptimalityCriterion,
+    GOptimalityCriterion,
+    AOptimalityCriterion,
+    EOptimalityCriterion,
+    DOptimalityCriterion,
+    IOptimalityCriterion,
     SpaceFillingCriterion,
-]
+)
 
 
 class DoEStrategy(Strategy):